feature: atualização.

+++ modificadoresdedados.exe/comoosperadoresagemdeformaexplicita.c
@@ -3,8 +3,8 @@ int main(){
     float  numero1 = 10.2;
     int numero2 = 10;
 
-    printf("numero1 > numero2: %d\n", numero1 > numero2);
-    printf("numero1 == numero2: %d\n", numero1 == numero2);
+    printf("numero1 > numero2: %d\n", (int)numero1 > numero2);
+    printf("numero1 == numero2: %d\n", (int)numero1 == numero2);
 
     return 0;
 }
